SJ-2056 Fixed delete of streams, services and providers. Fixed request get all instances

+++ sj-crud-rest/src/main/scala/com/bwsw/sj/crud/rest/api/SjProvidersApi.scala
@@ -70,15 +70,37 @@ trait SjProvidersApi extends Directives with SjCrudValidator {
             complete(HttpEntity(`application/json`, serializer.serialize(response)))
           } ~
           delete {
-            val provider = providerDAO.get(providerName)
-            var response: ProtocolResponse = null
-            if (provider != null) {
-              providerDAO.delete(providerName)
-              response = ProtocolResponse(200, Map("message" -> s"Provider '$providerName' has been deleted"))
-            } else {
-              response = ProtocolResponse(200, Map("message" -> s"Provider '$providerName' not found"))
+            val providers = serviceDAO.getAll.filter { service =>
+              service match {
+                case esService: ESService =>
+                  esService.provider.name.equals(providerName)
+                case zkService: ZKService =>
+                  zkService.provider.name.equals(providerName)
+                case aeroService: AerospikeService =>
+                  aeroService.provider.name.equals(providerName)
+                case cassService: CassandraService =>
+                  cassService.provider.name.equals(providerName)
+                case kfkService: KafkaService =>
+                  kfkService.provider.name.equals(providerName)
+                case tService: TStreamService =>
+                  tService.metadataProvider.name.equals(providerName) || tService.dataProvider.name.equals(providerName)
+                case _ =>
+                  false //todo redis and jdbc
+              }
             }
-            complete(HttpEntity(`application/json`, serializer.serialize(response)))
+            if (providers.isEmpty) {
+              val provider = providerDAO.get(providerName)
+              var response: ProtocolResponse = null
+              if (provider != null) {
+                providerDAO.delete(providerName)
+                response = ProtocolResponse(200, Map("message" -> s"Provider '$providerName' has been deleted"))
+              } else {
+                response = ProtocolResponse(200, Map("message" -> s"Provider '$providerName' not found"))
+              }
+              complete(HttpEntity(`application/json`, serializer.serialize(response)))
+            } else {
+              throw new BadRecordWithKey(s"Cannot delete provider $providerName. Provider usage in services", providerName)
+            }
           }
         } ~
         pathPrefix("connection") {
